backend/requirements.txt
# fastapi>=0.115.5
# uvicorn[standard]>=0.32.1
# redis==5.2.0
# pydantic==2.10.2
# pydantic-settings==2.6.1
# google-adk>=0.3.0
# anthropic==0.40.0
# prometheus-fastapi-instrumentator==7.0.0
# python-dotenv==1.0.1
# httpx==0.28.0


fastapi
uvicorn
redis
pydantic
pydantic-settings
google-adk
#anthropic
prometheus-fastapi-instrumentator
python-dotenv
httpx